+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- auth: Handle `UnicodeDecodeError` when loading JWT private key (#3).
+
 ## [1.0.2] - 2024-06-19
 
 ### Added